+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Disable resource link span ingestion. ([#2647](https://github.com/getsentry/relay/pull/2647))
 - Collect `http.decoded_response_content_length`. ([#2638](https://github.com/getsentry/relay/pull/2638))
 - Add TTID and TTFD tags to mobile spans. ([#2662](https://github.com/getsentry/relay/pull/2662))
+- Scrub all DB Core Data spans differently. ([#2686](https://github.com/getsentry/relay/pull/2686))
 
 ## 23.10.1